feat: étape 5 — occurrences croner + surface IPC complète + preview_schedule
- occurrence.rs : 5 prochaines dates (croner, secondes interdites),
  @-raccourcis dont @midnight, @reboot → vide ; timezone paramétrable,
  tests à timezone figée (§9)
- validation sémantique croner au create/update (99 9 * * * rejeté)
- commands Tauri §7 : list/get/create/update/delete/set_enabled/run_job/
  run_diagnostics/preview_schedule, spawn_blocking, AppState managé
- bindings.ts générés (tauri-specta) — régénérés par cargo test
- DECISIONS.md créé : RFC3339 en IPC, Interval u32, SchedulePreview

+++ src-tauri/src/model/mod.rs
@@ -49,7 +49,7 @@ pub enum ScheduleInfo {
     CronExpr(String),
     CalendarIntervals(Vec<CalendarEntry>),
     /// StartInterval en secondes.
-    Interval(u64),
+    Interval(u32),
 }
 
 /// Modèle commun en lecture (§3.1).
@@ -102,7 +102,7 @@ pub enum LaunchdCommand {
 #[serde(tag = "type", content = "value", rename_all = "camelCase")]
 pub enum LaunchdSchedule {
     CalendarIntervals(Vec<CalendarEntry>),
-    Interval(u64),
+    Interval(u32),
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
